# openapi.yaml
openapi: 3.0.3
info:
  title: Resort Backend API (compat + core)
  version: 1.0.0
  description: "Machine-readable contract for the Resort Backend used by the frontend. Includes compat `/api` endpoints, admin/internal headers, and upload specs."
servers:
  - url: "{API_BASE}"
    description: "Runtime server (use VITE_API_URL or REACT_APP_API_URL as API_BASE)"
    variables:
      API_BASE:
        default: "http://localhost:8000"
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    InternalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
  schemas:
...
    GalleryItem:
      type: object
      properties:
        id: { type: string }
        filename: { type: string }
        url: { type: string }
        category: { type: string }
        visible: { type: boolean }
paths:
...
  /experiences/:
    get:
      summary: List experiences
      responses: { '200': { description: OK } }
    post:
      requestBody: { content: { 'application/json': {} } }
      responses: { '201': { description: Created } }
  /experiences/{experience_id}:
    get:
      parameters: [{ name: experience_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    put:
      parameters: [{ name: experience_id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { 'application/json': {} } }
      responses: { '200': { description: Updated } }
    delete:
      parameters: [{ name: experience_id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted } }
...
  /menu-items/:
    get:
      summary: List menu items
      responses: { '200': { description: OK } }
    post:
      requestBody: { content: { 'application/json': {} } }
      responses: { '201': { description: Created } }
  /menu-items/{menu_item_id}:
    get:
      parameters: [{ name: menu_item_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    put:
      parameters: [{ name: menu_item_id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { 'application/json': {} } }
      responses: { '200': { description: Updated } }
    delete:
      parameters: [{ name: menu_item_id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted } }
  /gallery/:
    get:
      summary: List gallery
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: visible
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer }
        - name: skip
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Paged gallery
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/GalleryItem'
                  count: { type: integer }
    post:
      summary: Upload gallery item (admin)
      security: [{ AdminKey: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                imageUrl:
                  type: string
      responses:
        '201': { description: Created }
  /gallery/{item_id}:
    get:
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    put:
      security: [{ AdminKey: [] }]
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { 'application/json': {} } }
      responses: { '200': { description: Updated } }
    delete:
      security: [{ AdminKey: [] }]
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted } }
...
  /extra-beds/{item_id}:
    get:
      parameters: [{ name: item_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /extra-beds/for-accommodation/{accommodation_id}:
    get:
      parameters: [{ name: accommodation_id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /extra-beds/request:
    post:
      requestBody: { content: { 'application/json': {} } }
      responses: { '201': { description: Created } }
...
security: []